requirements.txt
Django==4.1.2
djangorestframework==3.14.0
django-cors-headers==3.13.0
python-dotenv==0.21.0
django-filter==23.1
firebase_admin==6.1.0
djangorestframework_simplejwt==5.3.0
postgres==4.0
django-storages==1.13.2
celery==5.2.7
redis==4.5.1
flower==1.2.0
django-ratelimit==4.0.0
requests==2.25.1
botocore==1.31.40
openai==0.28.0
Pillow==9.0.1